requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=/home/vatche/.work/ketabkala_invoice/requirements.txt /home/vatche/.work/ketabkala_invoice/requirements.in
#
aiohttp==3.8.4
    # via aiohttp-socks
aiohttp-socks==0.8.0
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via
    #   aiohttp
    #   python-socks
attrs==23.1.0
    # via aiohttp
beautifulsoup4==4.12.2
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
dnspython==2.3.0
    # via pymongo
exceptiongroup==1.1.2
    # via pytest
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
loguru==0.7.0
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.1
    # via pytest
peewee==3.16.2
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
pluggy==1.2.0
    # via pytest
psycopg2-binary==2.9.6
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
pymongo==4.4.0
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
pytest==7.4.0
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
python-dotenv==1.0.0
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
python-socks[asyncio]==2.3.0
    # via aiohttp-socks
requests==2.31.0
    # via -r /home/vatche/.work/ketabkala_invoice/requirements.in
soupsieve==2.4.1
    # via beautifulsoup4
tomli==2.0.1
    # via pytest
urllib3==2.0.3
    # via requests
yarl==1.9.2
    # via aiohttp
